Relocated "media-type-root" to under "workspace"
