+++ src/Umbraco.Web.UI.Client/src/packages/media/media-types/constants.ts
@@ -1,5 +1,4 @@
 export * from './entity-actions/constants.js';
-export * from './media-type-root/constants.js';
 export * from './paths.js';
 export * from './property-type/constants.js';
 export * from './repository/constants.js';
