+++ src/Umbraco.Web.UI.Client/src/packages/media/media-types/tree/tree-item-children/manifests.ts
@@ -3,7 +3,7 @@ import {
 	UMB_MEDIA_TYPE_FOLDER_ENTITY_TYPE,
 	UMB_MEDIA_TYPE_ROOT_ENTITY_TYPE,
 } from '../../entity.js';
-import { UMB_MEDIA_TYPE_ROOT_WORKSPACE_ALIAS } from '../../media-type-root/constants.js';
+import { UMB_MEDIA_TYPE_ROOT_WORKSPACE_ALIAS } from '../../workspace/media-type-root/constants.js';
 import { UMB_MEDIA_TYPE_FOLDER_WORKSPACE_ALIAS } from '../folder/constants.js';
 import { manifests as collectionManifests } from './collection/manifests.js';
