+++ src/Umbraco.Web.UI.Client/src/packages/media/media-types/workspace/constants.ts
@@ -1,2 +1,2 @@
-export const UMB_MEDIA_TYPE_WORKSPACE_ALIAS = 'Umb.Workspace.MediaType';
-export { UMB_MEDIA_TYPE_WORKSPACE_CONTEXT } from './media-type-workspace.context-token.js';
+export * from './media-type/constants.js';
+export * from './media-type-root/constants.js';
