+++ src/Umbraco.Web.UI.Client/src/packages/media/media-types/workspace/index.ts
@@ -1,1 +1,1 @@
-export * from './media-type-workspace.context-token.js';
+export * from './media-type/index.js';
