+++ src/Umbraco.Web.UI.Client/src/packages/media/media-types/workspace/manifests.ts
@@ -1,70 +1,4 @@
-import { UMB_MEDIA_TYPE_COMPOSITION_REPOSITORY_ALIAS } from '../constants.js';
-import { UMB_MEDIA_TYPE_WORKSPACE_ALIAS } from './constants.js';
-import { UMB_WORKSPACE_CONDITION_ALIAS, UmbSubmitWorkspaceAction } from '@umbraco-cms/backoffice/workspace';
+import { manifests as mediaTypeManifests } from './media-type/manifests.js';
+import { manifests as mediaTypeRootManifests } from './media-type-root/manifests.js';
 
-export const manifests: Array<UmbExtensionManifest> = [
-	{
-		type: 'workspace',
-		kind: 'routable',
-		alias: UMB_MEDIA_TYPE_WORKSPACE_ALIAS,
-		name: 'Media Type Workspace',
-		api: () => import('./media-type-workspace.context.js'),
-		meta: {
-			entityType: 'media-type',
-		},
-	},
-	{
-		type: 'workspaceView',
-		kind: 'contentTypeDesignEditor',
-		alias: 'Umb.WorkspaceView.MediaType.Design',
-		name: 'Media Type Workspace Design View',
-		meta: {
-			label: '#general_design',
-			pathname: 'design',
-			icon: 'icon-document-dashed-line',
-			compositionRepositoryAlias: UMB_MEDIA_TYPE_COMPOSITION_REPOSITORY_ALIAS,
-		},
-		conditions: [
-			{
-				alias: UMB_WORKSPACE_CONDITION_ALIAS,
-				match: UMB_MEDIA_TYPE_WORKSPACE_ALIAS,
-			},
-		],
-	},
-	{
-		type: 'workspaceView',
-		alias: 'Umb.WorkspaceView.MediaType.Structure',
-		name: 'Media Type Workspace Structure View',
-		element: () => import('./views/structure/media-type-workspace-view-structure.element.js'),
-		weight: 800,
-		meta: {
-			label: '#contentTypeEditor_structure',
-			pathname: 'structure',
-			icon: 'icon-mindmap',
-		},
-		conditions: [
-			{
-				alias: UMB_WORKSPACE_CONDITION_ALIAS,
-				match: UMB_MEDIA_TYPE_WORKSPACE_ALIAS,
-			},
-		],
-	},
-	{
-		type: 'workspaceAction',
-		kind: 'default',
-		alias: 'Umb.WorkspaceAction.MediaType.Save',
-		name: 'Save Media Type Workspace Action',
-		api: UmbSubmitWorkspaceAction,
-		meta: {
-			label: '#buttons_save',
-			look: 'primary',
-			color: 'positive',
-		},
-		conditions: [
-			{
-				alias: UMB_WORKSPACE_CONDITION_ALIAS,
-				match: UMB_MEDIA_TYPE_WORKSPACE_ALIAS,
-			},
-		],
-	},
-];
+export const manifests: Array<UmbExtensionManifest> = [...mediaTypeManifests, ...mediaTypeRootManifests];
