+++ src/Umbraco.Web.UI.Client/src/packages/media/media-types/workspace/media-type/constants.ts
@@ -1,0 +1,2 @@
+export const UMB_MEDIA_TYPE_WORKSPACE_ALIAS = 'Umb.Workspace.MediaType';
+export { UMB_MEDIA_TYPE_WORKSPACE_CONTEXT } from './media-type-workspace.context-token.js';
