+++ src/Umbraco.Web.UI.Client/src/packages/media/media-types/workspace/media-type/index.ts
@@ -1,0 +1,1 @@
+export * from './media-type-workspace.context-token.js';
