+++ src/Umbraco.Web.UI.Client/src/packages/media/media-types/workspace/media-type/media-type-workspace-editor.element.ts
