+++ src/Umbraco.Web.UI.Client/src/packages/media/media-types/workspace/media-type/media-type-workspace.context-token.ts
